Extracted matchers into testkit
